Fixes: se modifico getCategory y Category(model), se limitaron a 24 categorias hardcodeadas
