+++ client/src/components/Home/Home.jsx
@@ -20,7 +20,6 @@ function Home (){
     },[dispatch])
 
 
-
     return (
     <>
         <CreatedCarousel />
@@ -32,7 +31,7 @@ function Home (){
                                 <div class='card-container'>
                                     <Link to={`/products/${category.id}`} class='link'>
                                         <Card class='card'>
-                                            <Card.Img variant="top" src={category.img} />
+                                            <Card.Img variant="top" /*src={category.img}*/ />
                                             <Card.Body>
                                                 <Card.Title>{category.name}</Card.Title>
                                             </Card.Body>
