+++ client/src/components/Products/SearchBarProducts/SearchBarProducts.jsx
@@ -3,18 +3,25 @@ import Filters from "../Filters/Filters";
 import SearchBarProductsCards from "./SearchBarProductsCards";
 import style from "../Products.module.css";
 import { useLocation } from "react-router-dom";
-import { useSelector } from "react-redux";
+import { useDispatch, useSelector } from "react-redux";
+import { useEffect } from "react";
+import { getAllProducts } from "../../../redux/actions";
 
 function SearchBarProducts () {
     const location = useLocation()
+    const dispatch = useDispatch()
     const products = useSelector(state => state.products)
     const min = useSelector(state => state.minimumPrice);
     const max = useSelector(state => state.maximumPrice);
     const ratingFilterValue = useSelector(state => state.ratingFilterValue)
 
+    useEffect(() => {
+        dispatch(getAllProducts())
+    }, [dispatch])
+    const flatProducts = products.flat();
     const search = location.search.slice(8)
     
-    const searchProducts = products.filter((product) => product.name.includes(search))
+    const searchProducts = flatProducts.filter((product) => product.name.includes(search))
     const filteredProducts = searchProducts.filter((product) => {
         const matchSearch = product.name.includes(search);
         const matchPrice = (min && max) ? (product.price >= min && product.price <= max) :
