+++ client/src/redux/actionTypes.js
@@ -10,3 +10,4 @@ export const BETTER_QUALIFIED_FILTER = "BETTER_QUALIFIED_FILTER"
 export const ALL_FILTER = "ALL_FILTER"
 export const GET_ALL_CATEGORIES = 'GET_ALL_CATEGORIES'
 export const GET_ALL_PRODUCTS = "GET_ALL_PRODUCTS"
+export const GET_PRODUCT_BY_ID = "GET_PRODUCT_BY_ID"
