+++ client/src/redux/actions.js
@@ -1,4 +1,4 @@
-import { ADD_ORDER, ADD_PRODUCT,ADD_USER,REMOVE_ORDER,REMOVE_USER,REMOVE_PRODUCT,GET_ALL_CATEGORIES, MINIMUM_PRICE, MAXIMUM_PRICE, BETTER_QUALIFIED_FILTER, ALL_FILTER, GET_ALL_PRODUCTS } from "./actionTypes";
+import { ADD_ORDER, ADD_PRODUCT,ADD_USER,REMOVE_ORDER,REMOVE_USER,REMOVE_PRODUCT,GET_ALL_CATEGORIES, MINIMUM_PRICE, MAXIMUM_PRICE, BETTER_QUALIFIED_FILTER, ALL_FILTER, GET_ALL_PRODUCTS, GET_PRODUCT_BY_ID } from "./actionTypes";
 import axios from "axios";
 
 /////USER//////
@@ -160,6 +160,20 @@ export const getAllProducts = () => {
          })
       } catch (error) {
          console.log(error.message);
+      }
+   }
+}
+
+export const getProductById = (id) => {
+   return async function(dispatch){
+      try {
+         const response = await axios.get(`http://localhost:3001/PF/products/${id}`);
+         return dispatch({
+            type: GET_PRODUCT_BY_ID,
+            payload: response.data
+         })
+      } catch (error) {
+         console.log(error.message)
       }
    }
 }
