+++ server/src/controllers/getCategory.js
@@ -1,17 +1,16 @@
 const { Category } = require('../db');
 const axios = require("axios");
+const categoriesData = require("../data/categoriesData")
 
 
 const getCategory = async(req, res) => {
     try {
-        const respawns = await axios.get("https://api.mercadolibre.com/sites/MLA/categories")
-        const data = respawns.data;
+        const data = categoriesData;
         
         const category = await Category.findAll()
         if(category.length === 0){
             const categories = data.map((obj) => {
                 return {
-                    id: obj.id,
                     name: obj.name
                 }
             })
