+++ server/src/data/categoriesData.js
@@ -1,0 +1,100 @@
+categoriesData = [
+    {
+      name: "Accesorios para Vehículos",
+      img: "https://peru21.pe/resizer/Q-FJJsYBO0XDVoatA2evvlgq6FQ=/580x330/smart/filters:format(jpeg):quality(75)/arc-anglerfish-arc2-prod-elcomercio.s3.amazonaws.com/public/NPWZ3NUUK5BSBIK62Q33KTTZQI.jpg"
+    },
+    {
+      name: "Agro",
+      img: "https://agroapparg.com/img/arch-slider/intro_og.png"
+    },
+    {
+      name: "Alimentos y Bebidas",
+      img: "https://thefoodtech.com/wp-content/uploads/2022/05/alimentos-funcionales-principal.jpg"
+    },
+    {
+      name: "Animales y Mascotas",
+      img: "https://as01.epimg.net/diarioas/imagenes/2022/01/09/actualidad/1641723199_369382_1641723267_noticia_normal_recorte1.jpg"
+    },
+    {
+      name: "Antigüedades y Colecciones",
+      img: "https://http2.mlstatic.com/D_NQ_NP767510-MLA44229767080_122020-O.webp"
+    },
+    {
+      name: "Arte, Librería y Mercería",
+      img: "https://http2.mlstatic.com/D_NQ_NP657477-MLA43230937556_082020-O.webp"
+    },  
+    {
+      name: "Autos, Motos y Otros",
+      img: "https://www.auto-casa.com.ar/public/images/avisos/fotos/2876--07092021134523.jpg"
+    },
+    {
+      name: "Belleza y Cuidado Personal",
+      img: "https://diarioresponsable.com/images/CLARA_2021/unilever_belleza_positiva_dr.png"
+    },
+    {
+      name: "Celulares y Teléfonos",
+      img: "https://www.cronista.com/files/image/419/419139/61d3378d218ac.jpg"
+    },
+    {
+      name: "Computación",
+      img: "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcRQffZtTNgoiIjp82xHiyvKyymuWIMtu3Rz_Q&usqp=CAU"
+    },
+    {
+      name: "Consolas y Videojuegos",
+      img: "https://cnnespanol.cnn.com/wp-content/uploads/2021/12/211201164602-file-sony-playstation-5-microsoft-xbox-series-x-consoles-restricted-super-tease.jpg?quality=100&strip=info"
+    },
+    {
+      name: "Deportes y Fitness",
+      img: "https://2356021.fs1.hubspotusercontent-na1.net/hubfs/2356021/areas%20de%20un%20gimnasio.webp"
+    },
+    {
+      name: "Electrodomésticos y Aires Ac.",
+      img: "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcTdy50q1pV8ZPhkGevaIhAJ3hc-X_TPZj0gyg&usqp=CAU"
+    },
+    {
+      name: "Electrónica, Audio y Video",
+      img: "https://mundocdistribuciones.com/wp-content/uploads/2019/10/Electr%C3%B3nica-Audio-y-Video-1-2.jpg"
+    },
+    {
+      name: "Herramientas",
+      img: "https://definicion.de/wp-content/uploads/2010/06/herramienta-1.jpg"
+    },
+    {
+      name: "Hogar, Muebles y Jardín",
+      img: "https://media.admagazine.com/photos/618a625090c4ec9a52ca0f33/master/w_1600%2Cc_limit/81384.jpg"
+    },
+    {
+      name: "Instrumentos Musicles",
+      img: "https://www.elfinanciero.com.mx/resizer/KmztX7EtED6nwEhERiawZ8T8xD8=/400x267/filters:format(jpg):quality(70)/cloudfront-us-east-1.images.arcpublishing.com/elfinanciero/IEUM66UOW5ET3LSESPXRUULXU4.jpg"
+    },
+    {
+      name: "Joyas y Relojes",
+      img: "https://e00-expansion.uecdn.es/assets/multimedia/imagenes/2019/12/18/15766823152443.jpg"
+    },
+    {
+      name: "Juegos y Juguetes",
+      img: "https://www.clikisalud.net/wp-content/uploads/2016/10/shutterstock_320840780.jpg"
+    },
+    {
+      name: "Libros, Revistas y Comics",
+      img: "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcR3FOjTvChpVjaK2mUfCCqwQsamy7kBmj0mdw&usqp=CAU"
+    },  
+    {
+      name: "Música, Películas y Series",
+      img: "https://cloudfront-us-east-1.images.arcpublishing.com/artear/L4JXVA3U7WRCNLUT62YBMAJW24.jpg"
+    },
+    {
+      name: "Ropa y Accesorios",
+      img: "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQJFnsFFWii_iNSdUAEoz9lYu8P30-DI_LbZQ&usqp=CAU"
+    },
+    {
+      name: "Salud y Equipamiento Médico",
+      img: "https://concepto.de/wp-content/uploads/2013/08/salud-OMS-e1551914081412.jpg"
+    },
+    {
+      name: "Souvenirs, Cotillón y Fiestas",
+      img: "https://estaticos-cdn.prensaiberica.es/clip/2b02df77-1e52-4ca2-980b-606a0a153ab0_16-9-aspect-ratio_default_0.jpg"
+    }
+    ]
+
+    module.exports = categoriesData;
